% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioops.R
\name{\%is\%}
\alias{\%is\%}
\title{Strict identity comparison}
\usage{
a \%is\% b
}
\arguments{
\item{a}{First object.}

\item{b}{Second object.}
}
\value{
\code{TRUE} if identical, \code{FALSE} otherwise.
}
\description{
An infix operator for strict equality. Equivalent to \code{\link[base:identical]{base::identical()}}.
}
\examples{
1:3 \%is\% 1:3          # TRUE
1:3 \%is\% c(1, 2, 3)   # FALSE (integer vs double)
}
