% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioops.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not-in operator}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{A vector of values to test.}

\item{table}{A vector of values to test against.}
}
\value{
A logical vector.
}
\description{
Tests whether elements of \code{x} are \strong{not} present in \code{table}. Equivalent to \code{!(x \%in\% table)}.
NA behaviour follows base R semantics.
}
\examples{
c("A", "B", "C") \%nin\% c("B", "D")   # TRUE FALSE TRUE
1:5 \%nin\% c(2, 4)                    # TRUE FALSE TRUE FALSE TRUE
}
