% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioops.R
\name{\%p\%}
\alias{\%p\%}
\title{Paste two strings with a single space}
\usage{
lhs \%p\% rhs
}
\arguments{
\item{lhs}{A character vector.}

\item{rhs}{A character vector.}
}
\value{
A character vector of concatenated strings.
}
\description{
An infix operator for string concatenation with one space between \code{lhs} and \code{rhs}.
}
\examples{
"Hello" \%p\% "world"
c("hello", "good") \%p\% c("world", "morning")
}
