% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopalette.R
\name{list_palettes}
\alias{list_palettes}
\title{List Available Color Palettes}
\usage{
list_palettes(type = NULL, sort = TRUE)
}
\arguments{
\item{type}{Palette type(s) to filter: \code{"sequential"}, \code{"diverging"}, \code{"qualitative"}. Default NULL returns all.}

\item{sort}{Whether to sort by type, n_color, name. Default: TRUE.}
}
\value{
A \code{data.frame} with columns: \code{name}, \code{type}, \code{n_color}, \code{colors}.
}
\description{
Return a data.frame of all available palette metadata, optionally filtered by type.
}
\examples{
list_palettes()
list_palettes(type = "qualitative")
list_palettes(type = c("sequential", "diverging"))
}
