% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_chisq.R
\name{print.quick_chisq_result}
\alias{print.quick_chisq_result}
\title{Print Method for quick_chisq_result}
\usage{
\method{print}{quick_chisq_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{quick_chisq_result}}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for quick_chisq_result
}
