% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_cor.R
\name{print.quick_cor_result}
\alias{print.quick_cor_result}
\title{Print method for quick_cor_result}
\usage{
\method{print}{quick_cor_result}(x, ...)
}
\arguments{
\item{x}{A quick_cor_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for quick_cor_result
}
