% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_ttest.R
\name{print.quick_ttest_result}
\alias{print.quick_ttest_result}
\title{Print method for quick_ttest_result}
\usage{
\method{print}{quick_ttest_result}(x, ...)
}
\arguments{
\item{x}{A quick_ttest_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for quick_ttest_result
}
