% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind.R
\name{remind}
\alias{remind}
\title{Show usage tips for common R commands}
\usage{
remind(keyword = NULL)
}
\arguments{
\item{keyword}{A keyword like "glimpse" or "read_excel". If \code{NULL}, show all.}
}
\value{
Invisibly returns the matched keywords (character vector).
}
\description{
A helper to recall commonly used R functions with short examples.
}
\examples{
remind("glimpse")
remind()  # show all keywords
}
