% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_chisq.R
\name{summary.quick_chisq_result}
\alias{summary.quick_chisq_result}
\title{Summary Method for quick_chisq_result}
\usage{
\method{summary}{quick_chisq_result}(object, ...)
}
\arguments{
\item{object}{An object of class \code{quick_chisq_result}}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for quick_chisq_result
}
