% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_cor.R
\name{summary.quick_cor_result}
\alias{summary.quick_cor_result}
\title{Summary method for quick_cor_result}
\usage{
\method{summary}{quick_cor_result}(object, ...)
}
\arguments{
\item{object}{A quick_cor_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for quick_cor_result
}
