% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_nin_operator.R
\name{\%nin\%}
\alias{\%nin\%}
\title{❌ \verb{\%nin\%}: Not In Operator}
\usage{
... \%nin\% NA
}
\arguments{
\item{x}{A vector of values to test.}

\item{table}{A vector to match against.}
}
\value{
A logical vector indicating if elements of \code{x} are \strong{not in} \code{table}.
}
\description{
A binary operator to test if elements of the left-hand vector are \strong{not}
present in the right-hand vector. This is the negation of \code{\%in\%}, and is
equivalent to \code{!(x \%in\% y)}.
}
\examples{
c("A", "B", "C") \%nin\% c("B", "D")  # TRUE FALSE TRUE
1:5 \%nin\% c(2, 4)                   # TRUE FALSE TRUE FALSE TRUE
}
