% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_p_operator.R
\name{\%p\%}
\alias{\%p\%}
\title{Paste two strings with a space}
\usage{
lhs \%p\% rhs
}
\arguments{
\item{lhs}{A character vector on the left-hand side.}

\item{rhs}{A character vector on the right-hand side.}
}
\value{
A character vector, with elements from \code{lhs} and \code{rhs} pasted together using a space.
}
\description{
\verb{\%p\%} is a custom infix operator for string concatenation with a single space.
It’s inspired by the readability of \verb{\%>\%}, but meant for expressive text building.
}
\examples{
"Hello" \%p\% "world"
"Good" \%p\% "job"

}
