% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mirror.R
\name{set_mirror}
\alias{set_mirror}
\title{🔄 set_mirror: Set CRAN/Bioconductor Mirrors}
\usage{
set_mirror(repo = c("cran", "bioc", "all"), mirror = "cran")
}
\arguments{
\item{repo}{Character. "cran", "bioc", or "all".}

\item{mirror}{Character. Predefined mirror name or custom URL.}
}
\description{
Switch CRAN and/or Bioconductor mirrors for faster package installation.
}
\examples{
set_mirror("cran", "westlake")
set_mirror("bioc", "bioc_tuna")
set_mirror("all", "ustc")
}
